+++ projects/packages/blaze/src/class-blaze.php
@@ -9,10 +9,12 @@ namespace Automattic\Jetpack;
 
 use Automattic\Jetpack\Blaze\Dashboard as Blaze_Dashboard;
 use Automattic\Jetpack\Blaze\Dashboard_REST_Controller as Blaze_Dashboard_REST_Controller;
+use Automattic\Jetpack\Blaze\REST_Controller;
 use Automattic\Jetpack\Connection\Client;
 use Automattic\Jetpack\Connection\Initial_State as Connection_Initial_State;
 use Automattic\Jetpack\Connection\Manager as Jetpack_Connection;
 use Automattic\Jetpack\Status as Jetpack_Status;
+use Automattic\Jetpack\Status\Host;
 use Automattic\Jetpack\Sync\Settings as Sync_Settings;
 
 /**
@@ -46,8 +48,10 @@ class Blaze {
 		add_action( 'enqueue_block_editor_assets', array( __CLASS__, 'enqueue_block_editor_assets' ) );
 		// Add a Blaze Menu.
 		add_action( 'admin_menu', array( __CLASS__, 'enable_blaze_menu' ), 999 );
-		// Adds Blaze rest API
+		// Add Blaze dashboard app REST API endpoints.
 		add_action( 'rest_api_init', array( new Blaze_Dashboard_REST_Controller(), 'register_rest_routes' ) );
+		// Add general Blaze REST API endpoints.
+		add_action( 'rest_api_init', array( new REST_Controller(), 'register_rest_routes' ) );
 	}
 
 	/**
@@ -64,18 +68,27 @@ class Blaze {
 
 	/**
 	 * Is the wp-admin Dashboard enabled?
+	 * That dashboard is not available or necessary on WordPress.com sites.
 	 *
 	 * @return bool
 	 */
 	public static function is_dashboard_enabled() {
+		// Right now the dashboard is disabled by default.
+		$is_dashboard_enabled = false;
+
+		// On WordPress.com sites, the dashboard is not needed.
+		if ( ( new Host() )->is_wpcom_platform() ) {
+			$is_dashboard_enabled = false;
+		}
+
 		/**
 		 * Enable a wp-admin dashboard for Blaze campaign management.
 		 *
-		 * @since $$next-version$$
+		 * @since 0.7.0
 		 *
-		 * @param bool $should_enable Should the dashboard be enabled? False by default for now.
+		 * @param bool $should_enable Should the dashboard be enabled?
 		 */
-		return apply_filters( 'jetpack_blaze_dashboard_enable', false );
+		return apply_filters( 'jetpack_blaze_dashboard_enable', $is_dashboard_enabled );
 	}
 
 	/**
@@ -84,19 +97,33 @@ class Blaze {
 	 * @return void
 	 */
 	public static function enable_blaze_menu() {
-		if (
-			self::should_initialize()
-			&& self::is_dashboard_enabled()
-		) {
-			$blaze_dashboard = new Blaze_Dashboard();
-			$page_suffix     = add_submenu_page(
+		if ( ! self::should_initialize() ) {
+			return;
+		}
+
+		$blaze_dashboard = new Blaze_Dashboard();
+
+		if ( self::is_dashboard_enabled() ) {
+			$page_suffix = add_submenu_page(
 				'tools.php',
 				esc_attr__( 'Advertising', 'jetpack-blaze' ),
 				__( 'Advertising', 'jetpack-blaze' ),
 				'manage_options',
 				'advertising',
 				array( $blaze_dashboard, 'render' ),
-				100
+				1
+			);
+			add_action( 'load-' . $page_suffix, array( $blaze_dashboard, 'admin_init' ) );
+		} elseif ( ( new Host() )->is_wpcom_platform() ) {
+			$domain      = ( new Jetpack_Status() )->get_site_suffix();
+			$page_suffix = add_submenu_page(
+				'tools.php',
+				esc_attr__( 'Advertising', 'jetpack-blaze' ),
+				__( 'Advertising', 'jetpack-blaze' ),
+				'manage_options',
+				'https://wordpress.com/advertising/' . $domain,
+				null,
+				1
 			);
 			add_action( 'load-' . $page_suffix, array( $blaze_dashboard, 'admin_init' ) );
 		}