+++ projects/packages/my-jetpack/src/products/class-stats.php
@@ -7,8 +7,10 @@
 
 namespace Automattic\Jetpack\My_Jetpack\Products;
 
+use Automattic\Jetpack\Constants as Jetpack_Constants;
 use Automattic\Jetpack\My_Jetpack\Module_Product;
 use Automattic\Jetpack\My_Jetpack\Wpcom_Products;
+use Jetpack_Options;
 
 /**
  * Class responsible for handling the Jetpack Stats product
@@ -133,18 +135,38 @@ class Stats extends Module_Product {
 	 * @return boolean
 	 */
 	public static function has_required_plan() {
-		$purchases_data = Wpcom_Products::get_site_current_purchases();
-		if ( is_wp_error( $purchases_data ) ) {
-			return false;
-		}
-		if ( is_array( $purchases_data ) && ! empty( $purchases_data ) ) {
-			foreach ( $purchases_data as $purchase ) {
-				if ( 0 === strpos( $purchase->product_slug, 'jetpack_stats' ) ) {
-					return true;
+		// Check if paid stats plans have been enabled.
+		if ( Jetpack_Constants::is_true( 'JETPACK_PAID_STATS_ENABLED' ) ) {
+			$purchases_data = Wpcom_Products::get_site_current_purchases();
+			if ( is_wp_error( $purchases_data ) ) {
+				return false;
+			}
+			if ( is_array( $purchases_data ) && ! empty( $purchases_data ) ) {
+				foreach ( $purchases_data as $purchase ) {
+					if ( 0 === strpos( $purchase->product_slug, 'jetpack_stats' ) ) {
+						return true;
+					}
 				}
 			}
+			return false;
 		}
-		return false;
+
+		// Until the new paid stats plans roll out, no plan purchase is required for Jetpack Stats.
+		return true;
+	}
+
+	/**
+	 * Get the WordPress.com URL for purchasing Jetpack Stats for the current site.
+	 *
+	 * @return ?string
+	 */
+	public static function get_purchase_url() {
+		$blog_id = Jetpack_Options::get_option( 'id' );
+		// TODO: Handle unconnected sites without a defined blog_id. (Or check if we need to.)
+		// TODO: Remove the "stats/paid-stats" feature flag from the URL once paid stats has rolled out to the public.
+		// TODO: Consider adding a post-purchase redirect URL as a query string to the purchase URL.
+		// Appending get_manage_url() to the purchase URL would be a good option.
+		return sprintf( 'https://wordpress.com/stats/purchase/%d', $blog_id ) . '?flags=stats/paid-stats';
 	}
 
 	/**