+++ projects/plugins/jetpack/extensions/blocks/ai-assistant/extensions/ai-assistant/with-ai-assistant.tsx
@@ -2,21 +2,24 @@
  * External dependencies
  */
 import { BlockControls } from '@wordpress/block-editor';
-import { store as blockEditorStore } from '@wordpress/block-editor';
 import { createHigherOrderComponent } from '@wordpress/compose';
 import { useDispatch } from '@wordpress/data';
 import { useCallback, useState, useRef } from '@wordpress/element';
-import { store as noticesStore } from '@wordpress/notices';
 import React from 'react';
 /**
  * Internal dependencies
  */
 import AiAssistantDropdown, {
 	AiAssistantDropdownOnChangeOptionsArgProps,
+	KEY_ASK_AI_ASSISTANT,
 } from '../../components/ai-assistant-controls';
 import useSuggestionsFromAI, { SuggestionError } from '../../hooks/use-suggestions-from-ai';
 import { getPrompt } from '../../lib/prompt';
-import { getTextContentFromBlocks } from '../../lib/utils/block-content';
+import {
+	getRawTextFromHTML,
+	getTextContentFromSelectedBlocks,
+} from '../../lib/utils/block-content';
+import { transfromToAIAssistantBlock } from '../../transforms';
 /*
  * Types
  */
@@ -25,6 +28,13 @@ import type { PromptItemProps, PromptTypeProp } from '../../lib/prompt';
 type StoredPromptProps = {
 	messages: Array< PromptItemProps >;
 };
+
+/*
+ * An identifier to use on the extension error notices,
+ * so a existing notice with the same ID gets replaced
+ * by a new one, avoiding the stacking of notices.
+ */
+const AI_ASSISTANT_NOTICE_ID = 'ai-assistant';
 
 /*
  * Extend the withAIAssistant function of the block
@@ -38,13 +48,26 @@ export const withAIAssistant = createHigherOrderComponent(
 
 		const clientIdsRef = useRef< Array< string > >();
 
-		const { updateBlockAttributes, removeBlocks } = useDispatch( blockEditorStore );
-		const { createNotice } = useDispatch( noticesStore );
+		const { name: blockType } = props;
+
+		const { updateBlockAttributes, removeBlocks, replaceBlock } =
+			useDispatch( 'core/block-editor' );
+		const { createNotice } = useDispatch( 'core/notices' );
+
+		/*
+		 * Set exclude dropdown options.
+		 * - Exclude "Ask AI Assistant" for core/list-item block.
+		 */
+		const exclude = [];
+		if ( blockType === 'core/list-item' ) {
+			exclude.push( KEY_ASK_AI_ASSISTANT );
+		}
 
 		const showSuggestionError = useCallback(
 			( suggestionError: SuggestionError ) => {
 				createNotice( suggestionError.status, suggestionError.message, {
 					isDismissible: true,
+					id: AI_ASSISTANT_NOTICE_ID,
 				} );
 			},
 			[ createNotice ]
@@ -104,7 +127,7 @@ export const withAIAssistant = createHigherOrderComponent(
 			[ setStoredPrompt ]
 		);
 
-		const { request } = useSuggestionsFromAI( {
+		const { request, requestingState } = useSuggestionsFromAI( {
 			prompt: storedPrompt.messages,
 			onSuggestion: setContent,
 			onDone: updateStoredPrompt,
@@ -112,10 +135,10 @@ export const withAIAssistant = createHigherOrderComponent(
 			autoRequest: false,
 		} );
 
+		const { content, clientIds } = getTextContentFromSelectedBlocks();
+
 		const requestSuggestion = useCallback(
 			( promptType: PromptTypeProp, options: AiAssistantDropdownOnChangeOptionsArgProps ) => {
-				const { content, clientIds } = getTextContentFromBlocks();
-
 				/*
 				 * Store the selected clientIds when the user requests a suggestion.
 				 * The client Ids will be used to update the content of the block,
@@ -138,15 +161,27 @@ export const withAIAssistant = createHigherOrderComponent(
 					return freshPrompt;
 				} );
 			},
-			[ request ]
+			[ clientIds, content, request ]
 		);
+
+		const replaceWithAiAssistantBlock = useCallback( () => {
+			replaceBlock( props.clientId, transfromToAIAssistantBlock( { content }, blockType ) );
+		}, [ blockType, content, props.clientId, replaceBlock ] );
+
+		const rawContent = getRawTextFromHTML( props.attributes.content );
 
 		return (
 			<>
 				<BlockEdit { ...props } />
 
 				<BlockControls group="block">
-					<AiAssistantDropdown onChange={ requestSuggestion } />
+					<AiAssistantDropdown
+						requestingState={ requestingState }
+						disabled={ ! rawContent?.length }
+						onChange={ requestSuggestion }
+						onReplace={ replaceWithAiAssistantBlock }
+						exclude={ exclude }
+					/>
 				</BlockControls>
 			</>
 		);