+++ projects/plugins/jetpack/modules/tiled-gallery/tiled-gallery/tiled-gallery-shape.php
@@ -13,9 +13,23 @@ class Jetpack_Tiled_Gallery_Shape {
 	public static $shapes_used = array();
 
 	/**
+	 * The images to include in gallery.
+	 *
+	 * @var object[]
+	 */
+	protected $images;
+
+	/**
+	 * How many images are left after each row.
+	 *
+	 * @var int
+	 */
+	protected $images_left;
+
+	/**
 	 * Constructor class.
 	 *
-	 * @param object $images - the images.
+	 * @param object[] $images - the images.
 	 */
 	public function __construct( $images ) {
 		$this->images      = $images;