+++ projects/js-packages/components/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 ### This is a list detailing changes for the Jetpack RNA Components package releases.
 
+## [0.40.0] - 2023-06-26
+### Added
+- Add authentication to Zendesk chat. [#31339]
+
+## [0.39.0] - 2023-06-23
+### Added
+- Add component Pricing Slider for Stat pricing page.
+- Added config values to publish mirror repo
+- Auto-publish jetpack-components package to npm
+
 ## 0.38.1 - 2023-06-21
 ### Changed
 - Updated package dependencies. [#31468]
@@ -737,3 +747,6 @@
 
 ### Changed
 - Update node version requirement to 14.16.1
+
+[0.40.0]: https://github.com/Automattic/jetpack-components/compare/0.39.0...0.40.0
+[0.39.0]: https://github.com/Automattic/jetpack-components/compare/0.38.1...0.39.0